moved all css files out of component and added a new styles folder. Imports updated on all files
