+++ frontend/src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import "./index.css";
+import "./styles/global.css";
 import App from "./App.tsx";
 
 createRoot(document.getElementById("root")!).render(
