+++ frontend/src/styles/ProductDetailPage.module.css
