+++ frontend/src/styles/ProductResults.module.css
