+++ frontend/src/styles/SearchBar.module.css
