+++ frontend/src/styles/SearchPage.module.css
